added cargo feature control for numpy, polars and parquet
