+++ src/data_holder.rs
@@ -1,4 +1,7 @@
 pub mod arrow_helpers;
 pub mod channel_data;
 pub mod complex_arrow;
+#[cfg(feature = "numpy")]
+pub mod dtype;
+
 pub mod tensor_arrow;
