+++ src/data_holder/complex_arrow.rs
@@ -1,3 +1,4 @@
+//! complex number stored in primitive builders, fixedsizearraybuilder being too restricted
 use arrow::{
     array::{ArrayBuilder, BooleanBufferBuilder, PrimitiveArray, PrimitiveBuilder},
     buffer::{BooleanBuffer, MutableBuffer},
@@ -7,16 +8,21 @@ use arrow::{
 /// Complex struct
 #[derive(Debug)]
 pub struct ComplexArrow<T: ArrowPrimitiveType> {
+    /// The validity mask booolean buffer
     null_buffer_builder: Option<BooleanBuffer>,
+    /// the primitive builder
     values_builder: PrimitiveBuilder<T>,
+    /// the number of real and imaginary pairs in the array
     len: usize,
 }
 
-// Complex implementation
+/// Complex array implementation
 impl<T: ArrowPrimitiveType> ComplexArrow<T> {
+    /// Create new empty complex array
     pub fn new() -> Self {
         Self::with_capacity(1024)
     }
+    /// create new complex array with capacity
     pub fn with_capacity(capacity: usize) -> Self {
         Self {
             null_buffer_builder: None,
@@ -24,6 +30,7 @@ impl<T: ArrowPrimitiveType> ComplexArrow<T> {
             len: 0,
         }
     }
+    /// create new complex array from a MutableBuffer
     pub fn new_from_buffer(values_buffer: MutableBuffer) -> Self {
         let length = values_buffer.len() / 2;
         let values_builder = PrimitiveBuilder::new_from_buffer(values_buffer, None);
@@ -33,6 +40,7 @@ impl<T: ArrowPrimitiveType> ComplexArrow<T> {
             len: length,
         }
     }
+    /// Create a new complex array from a primitive builder and optionally its validity buffer
     pub fn new_from_primitive(
         primitive_builder: PrimitiveBuilder<T>,
         null_buffer: Option<&BooleanBuffer>,
@@ -48,30 +56,39 @@ impl<T: ArrowPrimitiveType> ComplexArrow<T> {
             len: length,
         }
     }
+    /// returns the mutable reference of the internal primitive builder array
     pub fn values(&mut self) -> &mut PrimitiveBuilder<T> {
         &mut self.values_builder
     }
+    /// retunrs the length of the complex array (number of real and imaginary pairs)
     pub fn len(&self) -> usize {
         self.len
     }
+    /// returns True if the complex array is empty
     pub fn is_empty(&self) -> bool {
         self.len() == 0
     }
+    /// returns the internal primitive builder as a slice
     pub fn values_slice(&self) -> &[T::Native] {
         self.values_builder.values_slice()
     }
+    /// returns the validity array
     pub fn nulls(&self) -> Option<&BooleanBuffer> {
         self.null_buffer_builder.as_ref()
     }
+    /// returns a finished cloned primitive array of the inner primitive builder
     pub fn finish_cloned(&self) -> PrimitiveArray<T> {
         self.values_builder.finish_cloned()
     }
+    /// returns a finished primitive array of the inner primitive builder
     pub fn finish(&mut self) -> PrimitiveArray<T> {
         self.values_builder.finish()
     }
+    /// overwrite the validity array
     pub fn set_validity(&mut self, mask: &mut BooleanBufferBuilder) {
         self.null_buffer_builder = Some(mask.finish());
     }
+    /// returns the optional validity array as a slice
     pub fn validity_slice(&self) -> Option<&[u8]> {
         self.null_buffer_builder.as_ref().map(|s| s.values())
     }
