+++ src/export/mod.rs
@@ -1,5 +1,7 @@
 //! Module to export mdf files to other file formats.
+#[cfg(feature = "numpy")]
 pub mod numpy;
+#[cfg(feature = "parquet")]
 pub mod parquet;
+#[cfg(feature = "polars")]
 pub mod polars;
-pub mod python_arrow_helpers;
