+++ src/main.rs
@@ -106,6 +106,7 @@ fn main() -> Result<(), Error> {
     }
 
     let parquet_compression = matches.get_one::<String>("parquet_compression");
+    #[cfg(feature = "parquet")]
     if let Some(file_name) = parquet_file_name {
         mdf_file
             .export_to_parquet(file_name, parquet_compression.map(|x| &**x))
